Refactor on searchChange function
