+++ frontend/src/utils/index.ts
@@ -1,5 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import { History, LocationState } from 'history';
 import { List, Set } from 'immutable';
+import queryString from 'query-string';
 import { DropdownItemProps } from 'semantic-ui-react';
 import {
   AggregateOptions,
@@ -162,3 +164,19 @@ export const getSelectOptionsFromSources = (sources: List<SourceMap>): DropdownI
       value: source.get('id'),
     }))
     .toJS();
+
+export const setQueryParams = (
+  value: string,
+  setPageNumber: React.Dispatch<React.SetStateAction<number>>,
+  setSearchQuery: React.Dispatch<React.SetStateAction<string>>,
+  history: History<LocationState>,
+  pathname: string,
+  search: string,
+): void => {
+  setSearchQuery(value || '');
+  setPageNumber(1);
+  const values = queryString.parse(search);
+  values.q = value || null;
+  values.page = '1';
+  history.push(`${pathname}?${queryString.stringify(values)}`);
+};
